update sentiment_analysis.py
Updated the sentiment_analysis.py file to keep the training and the testing code separate.
Also, the code is modified to accept command line parameter - "Train" or "Test" or "any garbage value"

In case of any other text other than - test and train, the code would exit. However, the use is free to enter the keyword in any case.